network refactor with dependency injection
