+++ app/src/main/java/com/example/rickandmorty/repository/Person.kt
@@ -1,4 +1,4 @@
-package com.example.rickandmorty.network
+package com.example.rickandmorty.repository
 
 data class Person(
     val id: Int,
