+++ app/src/main/java/com/example/rickandmorty/repository/Repository.kt
@@ -1,0 +1,5 @@
+package com.example.rickandmorty.repository
+
+interface Repository {
+    suspend fun getAllPersons() : List<Person>
+}
